git commits can't tame me
I've done pretty much everything. Dozens of breaking changes. There are some pretty neat features now. There is a classifier that adds tools to the context window based on training analysis. It works well. There is an autogen for the training data and it works pretty well, obviously things need human thinning but it gives you a good baseline. It allows you to drop a tool into the tools/ folder and it configures and sets the tool up so on the next load its ready to go. Pretty cool. The app also uses a SQLite database now because using json for file storage was untenable. There is a very busted Flask implementation that I have not gotten around to deleting and starting over. There are many supporting_files/ that are for generating tools. My hope is that someday you'll be able to drop a open-source API into a directory and run a "generate tool" command and it'll create a working tool you can use 5 minutes later. I'm actually pretty close to this if you can believe it. I can't figure out how to autonomously control Claude Code and feed it the right directions but I'll get there.

I'm reading the commit stuff and I didn't even have things like email_tool or maps_tool during the last commit. maps_tool can do neat stuff like "Which customer lives nearest to Lowe Mill" and it pulls the coordiantes of Lowe Mill and then references against the customers and returns the result in like 5 seconds total.

I'm sure there are other things I'm missing. I'm making this commit because I'm taking off a few days (till the 1st of the month) because I'm burnt out and its all I think about and I spent hundreds of dollars on the Anthropic API this month..... mostly the last one..